Eliminate 'comparison of ints of different signs' warning in GC_repeat_read (fix of commit 087187771)
Issue #558 (bdwgc).

* os_dep.c [LINUX_STACKBOTTOM || NEED_PROC_MAPS] (GC_repeat_read): Cast
num_read local variable to size_t in comparison.

+++ os_dep.c
@@ -103,7 +103,7 @@
     ssize_t num_read = 0;
 
     ASSERT_CANCEL_DISABLED();
-    while (num_read < count) {
+    while ((size_t)num_read < count) {
         ssize_t result = PROC_READ(fd, buf + num_read,
                                    count - (size_t)num_read);
